Starting full atom model

+++ Math/cMatrix44.h
@@ -9,14 +9,14 @@ class cMatrix44 {
 
 public:
 
-    double				m[4][4];
+    double	m[4][4];
     cMatrix44();
     cMatrix44(double mat[4][4]);
     cMatrix44(const cMatrix33 &rot, const cVector3 &shift);
-    cMatrix44(const float alpha, const float beta, const float R);
+    cMatrix44(const double alpha, const double beta, const double R);
 
-    void setGradAlpha(const float alpha, const float beta, const float R);
-    void setGradBeta(const float alpha, const float beta, const float R);
+    void setGradAlpha(const double alpha, const double beta, const double R);
+    void setGradBeta(const double alpha, const double beta, const double R);
 
     cMatrix44			operator*(const cMatrix44 &mat) const;
     cVector3			operator*(const cVector3 &vec) const;
